// src/app/visitas/models/interfaces/visitas.interface.ts
// Generated by https://quicktype.io

export interface IVisitaResponse {
    id: string;
    fecha: string;
    guiaNavigation: IGuia;
    horarioNavigation: IHorario;
    rutaNavigation: IRuta;
    visitanteNavigation: IVisitante;
    zonaNavigation: IZona;
}

export interface IVisitaRequest {
    fecha: string;
    guia: IGuia;
    horario: IHorario;
    ruta: IRuta;
    visitante: IVisitante;
    zona: IZona;
}

export interface IGuia {
    id: string;
    nombre: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    edad: number;
    email: string;
    telefono: string;
    direccion: string;
    lastLng?: string;
    lastLat?: string;
}

export interface IHorario {
    id:          string;
    horaInicio:  number;
    horaTermino: number;
    dia:         string;
    zona:        string;
}

export interface IRuta {
    id:          string;
    nombre:      string;
    descripcion: string;
    zona:        string;
}

export interface IVisitante {
    id:              string;
    nombre:          string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    edad:            number;
    email:           string;
    username:        string;
    telefono:        string;
    direccion:       string;
}

export interface IZona {
    id:          string;
    nombre:      string;
    descripcion: string;
    km2:         number;
}